# services/payments-svc/api/payments/v1/network_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.7.1
// - protoc             v6.33.0
// source: api/payments/v1/network.proto

package paymentsv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationNetworkOperationsServiceDistributeInterest = "/api.payments.v1.NetworkOperationsService/DistributeInterest"
const OperationNetworkOperationsServiceGetNetwork = "/api.payments.v1.NetworkOperationsService/GetNetwork"
const OperationNetworkOperationsServiceReconcile = "/api.payments.v1.NetworkOperationsService/Reconcile"
const OperationNetworkOperationsServiceRunNettingCycle = "/api.payments.v1.NetworkOperationsService/RunNettingCycle"
const OperationNetworkOperationsServiceSetClock = "/api.payments.v1.NetworkOperationsService/SetClock"

type NetworkOperationsServiceHTTPServer interface {
	// DistributeInterest DistributeInterest passes interest the Fed paid on the reserve account
	// to holders of settlement money through the accrual index, and pays each
	// member over Fedwire (412 while Fedwire is closed).
	DistributeInterest(context.Context, *DistributeInterestRequest) (*InterestDistribution, error)
	// GetNetwork GetNetwork returns members, the reserve account, the pool, cycles,
	// reconciliations and interest distributions.
	GetNetwork(context.Context, *GetNetworkRequest) (*Network, error)
	// Reconcile Reconcile attests the Fed's camt.052 balance for the reserve account
	// against the token's reserve pool.
	Reconcile(context.Context, *ReconcileRequest) (*Reconciliation, error)
	// RunNettingCycle RunNettingCycle plans the largest feasible set of queued obligations
	// and settles it through the plan book now.
	RunNettingCycle(context.Context, *RunNettingCycleRequest) (*NettingCycle, error)
	// SetClock SetClock moves the scenario clock (and the chain's) forward. Demo builds
	// only; a production build answers PERMISSION_DENIED.
	SetClock(context.Context, *SetClockRequest) (*SetClockResponse, error)
}

func RegisterNetworkOperationsServiceHTTPServer(s *http.Server, srv NetworkOperationsServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/v1/operator/network", _NetworkOperationsService_GetNetwork0_HTTP_Handler(srv))
	r.POST("/v1/operator/netting-cycles", _NetworkOperationsService_RunNettingCycle0_HTTP_Handler(srv))
	r.POST("/v1/operator/reconciliations", _NetworkOperationsService_Reconcile0_HTTP_Handler(srv))
	r.POST("/v1/operator/interest-distributions", _NetworkOperationsService_DistributeInterest0_HTTP_Handler(srv))
	r.POST("/v1/operator/clock", _NetworkOperationsService_SetClock0_HTTP_Handler(srv))
}

func _NetworkOperationsService_GetNetwork0_HTTP_Handler(srv NetworkOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetNetworkRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNetworkOperationsServiceGetNetwork)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetNetwork(ctx, req.(*GetNetworkRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Network)
		return ctx.Result(200, reply)
	}
}

func _NetworkOperationsService_RunNettingCycle0_HTTP_Handler(srv NetworkOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RunNettingCycleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNetworkOperationsServiceRunNettingCycle)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RunNettingCycle(ctx, req.(*RunNettingCycleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*NettingCycle)
		return ctx.Result(200, reply)
	}
}

func _NetworkOperationsService_Reconcile0_HTTP_Handler(srv NetworkOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ReconcileRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNetworkOperationsServiceReconcile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Reconcile(ctx, req.(*ReconcileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Reconciliation)
		return ctx.Result(200, reply)
	}
}

func _NetworkOperationsService_DistributeInterest0_HTTP_Handler(srv NetworkOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DistributeInterestRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNetworkOperationsServiceDistributeInterest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DistributeInterest(ctx, req.(*DistributeInterestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*InterestDistribution)
		return ctx.Result(200, reply)
	}
}

func _NetworkOperationsService_SetClock0_HTTP_Handler(srv NetworkOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SetClockRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNetworkOperationsServiceSetClock)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SetClock(ctx, req.(*SetClockRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SetClockResponse)
		return ctx.Result(200, reply)
	}
}

type NetworkOperationsServiceHTTPClient interface {
	DistributeInterest(ctx context.Context, req *DistributeInterestRequest, opts ...http.CallOption) (rsp *InterestDistribution, err error)
	GetNetwork(ctx context.Context, req *GetNetworkRequest, opts ...http.CallOption) (rsp *Network, err error)
	Reconcile(ctx context.Context, req *ReconcileRequest, opts ...http.CallOption) (rsp *Reconciliation, err error)
	RunNettingCycle(ctx context.Context, req *RunNettingCycleRequest, opts ...http.CallOption) (rsp *NettingCycle, err error)
	SetClock(ctx context.Context, req *SetClockRequest, opts ...http.CallOption) (rsp *SetClockResponse, err error)
}

type NetworkOperationsServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewNetworkOperationsServiceHTTPClient(client *http.Client) NetworkOperationsServiceHTTPClient {
	return &NetworkOperationsServiceHTTPClientImpl{client}
}

func (c *NetworkOperationsServiceHTTPClientImpl) DistributeInterest(ctx context.Context, in *DistributeInterestRequest, opts ...http.CallOption) (*InterestDistribution, error) {
	var out InterestDistribution
	pattern := "/v1/operator/interest-distributions"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNetworkOperationsServiceDistributeInterest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *NetworkOperationsServiceHTTPClientImpl) GetNetwork(ctx context.Context, in *GetNetworkRequest, opts ...http.CallOption) (*Network, error) {
	var out Network
	pattern := "/v1/operator/network"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationNetworkOperationsServiceGetNetwork))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *NetworkOperationsServiceHTTPClientImpl) Reconcile(ctx context.Context, in *ReconcileRequest, opts ...http.CallOption) (*Reconciliation, error) {
	var out Reconciliation
	pattern := "/v1/operator/reconciliations"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNetworkOperationsServiceReconcile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *NetworkOperationsServiceHTTPClientImpl) RunNettingCycle(ctx context.Context, in *RunNettingCycleRequest, opts ...http.CallOption) (*NettingCycle, error) {
	var out NettingCycle
	pattern := "/v1/operator/netting-cycles"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNetworkOperationsServiceRunNettingCycle))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *NetworkOperationsServiceHTTPClientImpl) SetClock(ctx context.Context, in *SetClockRequest, opts ...http.CallOption) (*SetClockResponse, error) {
	var out SetClockResponse
	pattern := "/v1/operator/clock"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNetworkOperationsServiceSetClock))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}