# services/payments-svc/api/payments/v1/notification_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.7.1
// - protoc             v6.33.0
// source: api/payments/v1/notification.proto

package paymentsv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationWebhookServiceCreateWebhook = "/api.payments.v1.WebhookService/CreateWebhook"
const OperationWebhookServiceDeleteWebhook = "/api.payments.v1.WebhookService/DeleteWebhook"
const OperationWebhookServiceListWebhookDeliveries = "/api.payments.v1.WebhookService/ListWebhookDeliveries"
const OperationWebhookServiceListWebhooks = "/api.payments.v1.WebhookService/ListWebhooks"

type WebhookServiceHTTPServer interface {
	// CreateWebhook CreateWebhook registers an https endpoint; the secret is returned once.
	CreateWebhook(context.Context, *CreateWebhookRequest) (*Webhook, error)
	// DeleteWebhook DeleteWebhook removes an endpoint; its pending deliveries are dropped.
	DeleteWebhook(context.Context, *DeleteWebhookRequest) (*DeleteWebhookResponse, error)
	// ListWebhookDeliveries ListWebhookDeliveries is an endpoint's delivery log.
	ListWebhookDeliveries(context.Context, *ListWebhookDeliveriesRequest) (*ListWebhookDeliveriesResponse, error)
	// ListWebhooks ListWebhooks lists the organization's endpoints, without secrets.
	ListWebhooks(context.Context, *ListWebhooksRequest) (*ListWebhooksResponse, error)
}

func RegisterWebhookServiceHTTPServer(s *http.Server, srv WebhookServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/webhooks", _WebhookService_CreateWebhook0_HTTP_Handler(srv))
	r.GET("/v1/webhooks", _WebhookService_ListWebhooks0_HTTP_Handler(srv))
	r.DELETE("/v1/webhooks/{id}", _WebhookService_DeleteWebhook0_HTTP_Handler(srv))
	r.GET("/v1/webhooks/{id}/deliveries", _WebhookService_ListWebhookDeliveries0_HTTP_Handler(srv))
}

func _WebhookService_CreateWebhook0_HTTP_Handler(srv WebhookServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateWebhookRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWebhookServiceCreateWebhook)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateWebhook(ctx, req.(*CreateWebhookRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Webhook)
		return ctx.Result(200, reply)
	}
}

func _WebhookService_ListWebhooks0_HTTP_Handler(srv WebhookServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListWebhooksRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWebhookServiceListWebhooks)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListWebhooks(ctx, req.(*ListWebhooksRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListWebhooksResponse)
		return ctx.Result(200, reply)
	}
}

func _WebhookService_DeleteWebhook0_HTTP_Handler(srv WebhookServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteWebhookRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWebhookServiceDeleteWebhook)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteWebhook(ctx, req.(*DeleteWebhookRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteWebhookResponse)
		return ctx.Result(200, reply)
	}
}

func _WebhookService_ListWebhookDeliveries0_HTTP_Handler(srv WebhookServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListWebhookDeliveriesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWebhookServiceListWebhookDeliveries)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListWebhookDeliveries(ctx, req.(*ListWebhookDeliveriesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListWebhookDeliveriesResponse)
		return ctx.Result(200, reply)
	}
}

type WebhookServiceHTTPClient interface {
	CreateWebhook(ctx context.Context, req *CreateWebhookRequest, opts ...http.CallOption) (rsp *Webhook, err error)
	DeleteWebhook(ctx context.Context, req *DeleteWebhookRequest, opts ...http.CallOption) (rsp *DeleteWebhookResponse, err error)
	ListWebhookDeliveries(ctx context.Context, req *ListWebhookDeliveriesRequest, opts ...http.CallOption) (rsp *ListWebhookDeliveriesResponse, err error)
	ListWebhooks(ctx context.Context, req *ListWebhooksRequest, opts ...http.CallOption) (rsp *ListWebhooksResponse, err error)
}

type WebhookServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewWebhookServiceHTTPClient(client *http.Client) WebhookServiceHTTPClient {
	return &WebhookServiceHTTPClientImpl{client}
}

func (c *WebhookServiceHTTPClientImpl) CreateWebhook(ctx context.Context, in *CreateWebhookRequest, opts ...http.CallOption) (*Webhook, error) {
	var out Webhook
	pattern := "/v1/webhooks"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationWebhookServiceCreateWebhook))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *WebhookServiceHTTPClientImpl) DeleteWebhook(ctx context.Context, in *DeleteWebhookRequest, opts ...http.CallOption) (*DeleteWebhookResponse, error) {
	var out DeleteWebhookResponse
	pattern := "/v1/webhooks/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWebhookServiceDeleteWebhook))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *WebhookServiceHTTPClientImpl) ListWebhookDeliveries(ctx context.Context, in *ListWebhookDeliveriesRequest, opts ...http.CallOption) (*ListWebhookDeliveriesResponse, error) {
	var out ListWebhookDeliveriesResponse
	pattern := "/v1/webhooks/{id}/deliveries"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWebhookServiceListWebhookDeliveries))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *WebhookServiceHTTPClientImpl) ListWebhooks(ctx context.Context, in *ListWebhooksRequest, opts ...http.CallOption) (*ListWebhooksResponse, error) {
	var out ListWebhooksResponse
	pattern := "/v1/webhooks"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWebhookServiceListWebhooks))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}