# services/payments-svc/internal/server/openapi.yaml
...
openapi: 3.0.3
info:
    title: Reserve interbank settlement payment hub API
    version: 1.0.0
paths:
...
    /v1/operator/clock:
        post:
            tags:
                - NetworkOperationsService
            description: |-
                SetClock moves the scenario clock (and the chain's) forward. Demo builds
                 only; a production build answers PERMISSION_DENIED.
            operationId: NetworkOperationsService_SetClock
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.SetClockRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.SetClockResponse'
    /v1/operator/interest-distributions:
        post:
            tags:
                - NetworkOperationsService
            description: |-
                DistributeInterest passes interest the Fed paid on the reserve account
                 to holders of settlement money through the accrual index, and pays each
                 member over Fedwire (412 while Fedwire is closed).
            operationId: NetworkOperationsService_DistributeInterest
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.DistributeInterestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.InterestDistribution'
    /v1/operator/netting-cycles:
        post:
            tags:
                - NetworkOperationsService
            description: |-
                RunNettingCycle plans the largest feasible set of queued obligations
                 and settles it through the plan book now.
            operationId: NetworkOperationsService_RunNettingCycle
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.RunNettingCycleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.NettingCycle'
...
    /v1/operator/reconciliations:
        post:
            tags:
                - NetworkOperationsService
            description: |-
                Reconcile attests the Fed's camt.052 balance for the reserve account
                 against the token's reserve pool.
            operationId: NetworkOperationsService_Reconcile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.ReconcileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Reconciliation'
...
    /v1/webhooks:
        get:
            tags:
                - WebhookService
            description: ListWebhooks lists the organization's endpoints, without secrets.
            operationId: WebhookService_ListWebhooks
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.ListWebhooksResponse'
        post:
            tags:
                - WebhookService
            description: CreateWebhook registers an https endpoint; the secret is returned once.
            operationId: WebhookService_CreateWebhook
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.CreateWebhookRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Webhook'
    /v1/webhooks/{id}:
        delete:
            tags:
                - WebhookService
            description: DeleteWebhook removes an endpoint; its pending deliveries are dropped.
            operationId: WebhookService_DeleteWebhook
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.DeleteWebhookResponse'
...
components:
    schemas:
...
        api.payments.v1.CreateWebhookRequest:
            type: object
            properties:
                url:
                    type: string
...
        api.payments.v1.DeleteWebhookResponse:
            type: object
            properties: {}
        api.payments.v1.DistributeInterestRequest:
            type: object
            properties:
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
...
        api.payments.v1.InterestDistribution:
            type: object
            properties:
                id:
                    type: string
                at:
                    type: string
                    format: date-time
                runBy:
                    type: string
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                paid:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                retained:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                shares:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/api.payments.v1.Money'
...
        api.payments.v1.ListWebhooksResponse:
            type: object
            properties:
                webhooks:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.Webhook'
...
        api.payments.v1.Money:
            type: object
            properties:
                amount:
                    type: string
                currency:
                    type: string
            description: |-
                Money is an amount in a currency. The amount is a decimal string with at
                 most two fraction digits ("25000000.00"), never a float.
        api.payments.v1.NettingCycle:
            type: object
            properties:
                id:
                    type: string
                cycleRef:
                    type: string
                at:
                    type: string
                    format: date-time
                trigger:
                    type: string
                discharged:
                    type: integer
                    format: int32
                deferred:
                    type: integer
                    format: int32
                gross:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                net:
                    $ref: '#/components/schemas/api.payments.v1.Money'
...
        api.payments.v1.ReconcileRequest:
            type: object
            properties: {}
        api.payments.v1.Reconciliation:
            type: object
            properties:
                id:
                    type: string
                at:
                    type: string
                    format: date-time
                fedBalance:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                reservePool:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                reconciliationBreak:
                    type: boolean
                invariantsHold:
                    type: boolean
...
        api.payments.v1.RunNettingCycleRequest:
            type: object
            properties: {}
        api.payments.v1.SetClockRequest:
            type: object
            properties:
                time:
                    type: string
                    format: date-time
        api.payments.v1.SetClockResponse:
            type: object
            properties:
                now:
                    type: string
                    format: date-time
...
        api.payments.v1.Webhook:
            type: object
            properties:
                id:
                    type: string
                url:
                    type: string
                secret:
                    type: string
                createdAt:
                    type: string
                    format: date-time
...